les3

+++ les02/pe2_4.py
@@ -1,3 +1,3 @@
 uurloon = eval(input('Wat is je uurloon: '))
 aantalUur = eval(input('Hoeveel uur heb je gewerkt: '))
-print(str(aantalUur) + ' uur werken levert ' + str(float(uurloon * aantalUur)) + ' Euro op')
+print(str(aantalUur) + ' uur werken levert ' + str(uurloon * aantalUur) + ' Euro op')
